% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_chr.R
\name{mapa_chr}
\alias{mapa_chr}
\title{mapa_chr}
\usage{
mapa_chr(cr = c(30.4, 19.7, 23.5, 18.6, 27), col_cr = "grey",
  unidad = "cM")
}
\arguments{
\item{cr}{vector}

\item{col_cr}{color}

\item{unidad}{cM}
}
\description{
mapa_chr
}
